+++ lib/country_phone_picker.dart
@@ -1,7 +1,8 @@
 export 'src/bottom_sheet_config.dart';
 export 'src/country_model.dart';
 export 'src/country_phone_picker.dart';
-export 'src/country_picker_codes.dart' show findCountryByIsoCode;
+export 'src/country_picker_codes.dart'
+    show findCountryByIsoCode, countriesForIsoCodes;
 export 'src/country_phone_picker_bottom_sheet.dart';
 export 'src/country_picker_localizations.dart';
 export 'src/search_config.dart';